+++ src/optional_tasks/task.py
@@ -2,16 +2,16 @@ from optional_tasks.times import get_unix_day_start_time
 
 
 class Task:
-    def __init__(self, id: int, name: str, difficulty: int, tags: set, completions: list = []) -> None:
+    def __init__(self, id: int, description: str, priority: int, tags: set, completions: list = []) -> None:
         self.id = id
-        self.name = name
-        self.difficulty = difficulty
+        self.description = description
+        self.priority = priority
         self.tags = tags
         self.completions = completions
 
     @classmethod
     def from_dict(cls, data: dict):
-        return cls(data['id'], data['name'], data['difficulty'], set(data['tags']), data['completions'])
+        return cls(data['id'], data['description'], data['priority'], set(data['tags']), data['completions'])
 
     def __str__(self):
         return self.__dict__.__str__()