+++ src/optional_tasks/tasks.py
@@ -22,8 +22,8 @@ class Tasks:
         if is_short_header:
             self.headers = {
                 'id': 'id',
-                'name': 'name',
-                'difficulty': 'd',
+                'description': 'description',
+                'priority': 'p',
                 'tags': 'tags',
                 'count': 'c',
                 'score': 's',
@@ -34,8 +34,8 @@ class Tasks:
         else:
             self.headers = {
                 'id': 'id',
-                'name': 'name',
-                'difficulty': 'difficulty',
+                'description': 'description',
+                'priority': 'priority',
                 'tags': 'tags',
                 'count': 'count',
                 'score': 'score',
@@ -45,37 +45,42 @@ class Tasks:
             }
         self.lens = self.__count_lengths()
 
-    def add(self, name: str, difficulty: int, tags: set):
+    def add(self, description: str, priority: int, tags):
         ids = [task.id for task in self.tasks if task.id != None] or [-1]
-        names: list = [task.name for task in self.tasks]
-        if name not in names:
-            self.tasks.append(Task(max(ids) + 1, name, difficulty or 1, tags))
+        descriptions: list = [task.description for task in self.tasks]
+        if description not in descriptions:
+            self.tasks.append(Task(max(ids) + 1, description, priority, tags))
         else:
             raise exceptions.NameAlreadyExistError(
-                'Task with this name already exists')
+                'Task with this description already exists')
 
-    def edit(self, id: int, name, difficulty: int, tags: set):
+    def edit(self, id, description, priority: int, tags):
         target_task = None
         for task in self.tasks:
             if task.id == id:
                 target_task = task
                 self.tasks.remove(task)
-        names: list = [task.name for task in self.tasks]
-        if name not in names and target_task:
+        descriptions: list = [task.description for task in self.tasks]
+        if description not in descriptions and target_task:
             self.tasks.append(Task(target_task.id,
-                                   name or target_task.name,
-                                   difficulty or target_task.difficulty,
+                                   description or target_task.description,
+                                   priority or target_task.priority,
                                    tags or target_task.tags,
                                    target_task.completions))
         else:
             self.tasks.append(target_task)
             raise exceptions.NameAlreadyExistError(
-                'Task with this name already exists')
+                'Task with this description already exists')
 
     def complete(self, id):
         for task in self.tasks:
             if task.id == id:
                 task.completions.append(get_current_unix_time())
+
+    def remove(self, id):
+        for task in self.tasks:
+            if task.id == id:
+                self.tasks.remove(task)
 
     def __del__(self):
         self.tasks.sort(key=lambda x: x.id, reverse=False)
@@ -84,21 +89,21 @@ class Tasks:
         files.save_json([task.__dict__ for task in self.tasks], self.file)
 
     def __sort(self, sort: str):
-        if (sort == 'name'):
-            self.tasks.sort(key=lambda x: x.name, reverse=False)
-        elif (sort == 'difficulty'):
-            self.tasks.sort(key=lambda x: x.difficulty, reverse=False)
+        if (sort == 'description'):
+            self.tasks.sort(key=lambda x: x.description, reverse=False)
+        elif (sort == 'priority'):
+            self.tasks.sort(key=lambda x: x.priority, reverse=False)
         elif (sort == 'count'):
             self.tasks.sort(key=lambda x: len(x.completions), reverse=False)
         elif (sort == 'score'):
             self.tasks.sort(key=lambda x: len(x.completions)
-                            * x.difficulty, reverse=False)
+                            * x.priority, reverse=False)
         elif (sort == 'tcount'):
             self.tasks.sort(key=lambda x: len(
                 x.get_today_completions()), reverse=False)
         elif (sort == 'tscore'):
             self.tasks.sort(key=lambda x: len(x.get_today_completions())
-                            * x.difficulty, reverse=False)
+                            * x.priority, reverse=False)
         elif (sort == 'date'):
             self.tasks.sort(key=lambda x: max(
                 x.completions or [0]), reverse=False)
@@ -111,36 +116,36 @@ class Tasks:
         if self.tasks:
             lens['id'] = max(max([len(str(task.id))
                              for task in self.tasks]), len(self.headers['id']))
-            lens['name'] = max(max([len(task.name)
-                               for task in self.tasks]), len(self.headers['name']))
+            lens['description'] = max(max([len(task.description)
+                                           for task in self.tasks]), len(self.headers['description']))
             lens['tags'] = max(max([len(task.get_tags())
                                for task in self.tasks]), len(self.headers['tags']))
             lens['count'] = max(
                 max([len(str(len(task.completions))) for task in self.tasks]), len(self.headers['count']))
             lens['score'] = max(
-                max([len(str(len(task.completions)*task.difficulty)) for task in self.tasks]), len(self.headers['score']))
+                max([len(str(len(task.completions)*task.priority)) for task in self.tasks]), len(self.headers['score']))
             lens['today count'] = max(
                 max([len(str(len(task.get_today_completions()))) for task in self.tasks if task]), len(self.headers['today count']))
             lens['today score'] = max(
-                max([len(str(len(task.get_today_completions())*task.difficulty)) for task in self.tasks]), len(self.headers['today score']))
+                max([len(str(len(task.get_today_completions())*task.priority)) for task in self.tasks]), len(self.headers['today score']))
         else:
             lens['id'] = 2  # 'id'
-            lens['name'] = len(self.headers['name'])  # 'name'
+            lens['description'] = len(
+                self.headers['description'])  # 'description'
             lens['tags'] = len(self.headers['tags'])  # 'tags'
             lens['count'] = len(self.headers['count'])  # 'count'
             lens['score'] = len(self.headers['score'])  # 'score'
             lens['today count'] = len(self.headers['today count'])  # 'count'
             lens['today score'] = len(self.headers['today score'])  # 'score'
-        lens['difficulty'] = max(
-            len(self.headers['difficulty']), 2)   # 'difficulty'
+        lens['priority'] = len(self.headers['priority'])   # 'priority'
         lens['last completion date'] = max(
             len(self.headers['last completion date']), 19)
         return lens
 
     def __print_header(self):
         output = (f'{Colors.UNDERLINE}{self.headers["id"]:<{self.lens["id"]}}{Colors.ENDC} '
-                  f'{Colors.UNDERLINE}{self.headers["name"]:<{self.lens["name"]}}{Colors.ENDC} '
-                  f'{Colors.UNDERLINE}{self.headers["difficulty"]:<{self.lens["difficulty"]}}{Colors.ENDC} '
+                  f'{Colors.UNDERLINE}{self.headers["description"]:<{self.lens["description"]}}{Colors.ENDC} '
+                  f'{Colors.UNDERLINE}{self.headers["priority"]:<{self.lens["priority"]}}{Colors.ENDC} '
                   f'{Colors.UNDERLINE}{self.headers["tags"]:<{self.lens["tags"]}}{Colors.ENDC} '
                   f'{Colors.UNDERLINE}{self.headers["count"]:<{self.lens["count"]}}{Colors.ENDC} '
                   f'{Colors.UNDERLINE}{self.headers["score"]:<{self.lens["score"]}}{Colors.ENDC} '
@@ -152,13 +157,13 @@ class Tasks:
     def __print_row(self, task: Task):
         last_completion_date = max(task.completions or [0])
         output = (f'{task.id:<{self.lens["id"]}} '
-                  f'{task.name:<{self.lens["name"]}} '
-                  f'{task.difficulty:<{self.lens["difficulty"]}} '
+                  f'{task.description:<{self.lens["description"]}} '
+                  f'{task.priority:<{self.lens["priority"]}} '
                   f'{task.get_tags():<{self.lens["tags"]}} '
                   f'{len(task.completions):<{self.lens["count"]}} '
-                  f'{len(task.completions)*task.difficulty:<{self.lens["score"]}} '
+                  f'{len(task.completions)*task.priority:<{self.lens["score"]}} '
                   f'{len(task.get_today_completions()):<{self.lens["today count"]}} '
-                  f'{len(task.get_today_completions())*task.difficulty:<{self.lens["today score"]}} ')
+                  f'{len(task.get_today_completions())*task.priority:<{self.lens["today score"]}} ')
         if last_completion_date == 0:
             output += f'{Colors.FAIL}{"never":<{self.lens["last completion date"]}}{Colors.ENDC}'
         elif last_completion_date > get_unix_day_start_time():
@@ -177,7 +182,7 @@ class Tasks:
                 if unix_day_start <= completion < unix_day_end:
                     if unix_day_start not in date_score:
                         date_score[unix_day_start] = 0
-                    date_score[unix_day_start] += task.difficulty
+                    date_score[unix_day_start] += task.priority
         try:
             average_score = sum(date_score.values())/len(date_score.values())
         except ZeroDivisionError as e:
@@ -193,14 +198,14 @@ class Tasks:
     def print(self, group: str, sort: str):
         self.__sort(sort)
         self.__print_header()
-        if (group == 'difficulty'):
+        if (group == 'priority'):
             difficulties = set()
             for task in self.tasks:
-                difficulties.add(task.difficulty)
+                difficulties.add(task.priority)
             difficulties = sorted(difficulties)
-            for difficulty in difficulties:
-                print(f'difficulty: {difficulty}')
-                for task in [task for task in self.tasks if task.difficulty == difficulty]:
+            for priority in difficulties:
+                print(f'priority: {priority}')
+                for task in [task for task in self.tasks if task.priority == priority]:
                     self.__print_row(task)
                 print()
         elif (group == 'tags'):
@@ -211,7 +216,7 @@ class Tasks:
                         tags[tag] = 0
             for tag in tags:
                 for task in [task for task in self.tasks if tag in task.tags]:
-                    tags[tag] += len(task.completions) * task.difficulty
+                    tags[tag] += len(task.completions) * task.priority
             tags = dict(
                 sorted(tags.items(), key=lambda item: item[1], reverse=True))
             for tag in tags: