+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [0.7.0] - 2021-07-19
+### Added
+- Class Args for parsing arguments #7.
+- Deleting task #6.
+
+### Changed
+- Task name rename to description.
+- Task difficulty rename to priority.
+
 ## [0.6.0] - 2021-07-15
 ### Added
 - Sorting by today task complitions and today task score.